02_Embeddings_and_RAG/pyproject.toml
[project]
name = "02-embeddings-and-rag"
version = "0.1.0"
description = "Embeddings and RAG"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ipykernel>=6.29.5",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.0",
    "numpy>=1.26.0",
    "openai>=1.59.7",
    "pandas>=2.2.3",
    "plotly>=5.24.1",
    "python-dotenv>=1.0.1",
    "scikit-learn>=1.6.1",
    "scipy>=1.15.1",
    "PyPDF2>=3.0.0",
    "requests>=2.31.0",
    "faiss-cpu>=1.7.4",
    "annoy>=1.17.3",
    "graphviz>=0.21",
    "xai-sdk>=1.1.0",
]
